Update patch set 4
Patch Set 4:

> FAIL: Some checks failed. See below.
 > 
 > 
 > === git_master/gce_x86_phone-userdebug_fastbuild3c_linux ===
 > 
 > CONFLICT: Your changes failed to cherry-pick against build 3749927
 > (could not apply fa17f9f... Update FreeType to e432ebf2 (2.7.1 +
 > patches)). See logs/git.log in https://android-build.googleplex.com/builds/pending/P3882028/gce_x86_phone-userdebug_fastbuild3c_linux/latest.
 
A conflict with git-master is due to a cherry-pick  from the upstream to improve bitmap font metric retrieval. 

I'll make a CL against git-master and merge it to git-master. Once it's merged to git-master, I'll add "Merged-In: <git-master cl's ID>"  to this CL's description. 

That way, this CL will not be merged back to git-master.  And, aosp-master will be in sync with git-master for external/freetype.

Patch-set: 4
Reviewer: jungshik <1060643@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0