Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > Please update README.version as well.
> > > 
> > > just out of curiosity, what is the motivation of the upgrading?
> > > As far as I know, 2.9.1 is the latest stable version.
> > > Is there anything critical issues?
> > 
> > There's no critical issue. I'm building a tool to upgrade all libraries in external/ in general.
> > 
> > It makes sense to follow stable version. In that case, it is better to write the version git tag in METADATA. So that my tool will only generate change when there's a new version. I'll send a cl soon.
> 
> Actually current version is beyond 2.9.1. See b/113673743. Can you confirm whether you want to follow ToT or stable version? (Or don't care but want to manually upgrade?)
> 
> If set to ToT, my tool will send upgrade if there's any new change and we didn't upgrade in last 1 month.

Hi Haibo

We have a dilemma here since everytime we update freetype we take the risk of breaking something in android text which has impacts users to the largest end. 

We do not stay on stable version but as written in README we try to follow Chrome. A lot of times we forget to update too. However, for example now just a few months before Q release we also would not like to make changes to FreeType or other native libraries such as Harfbuzz. 

As long as the tool creates the CL and we accept or reject them I think it should be fine to keep ToT. 

ToT might have both security issues and also font parsing/rendering issues. Therefore when we create a patch, and before merging it we make sure that automated tests pass, and also we flash and see if overall things looks fine as a smoke test.

Can you guide us about what you think? Are there any other libraries with similar requirements?

Patch-set: 4
Reviewer: Gerrit User 1106808 <1106808@85c56323-6fa9-3386-8a01-6480fb634889>