Use 'private: true' for the LLDNK-private library
We used 'vendor_available: false' to define VNDK-private libraries.
However, 'vendor_available' is originally used to create vendor
variants that is available to vendor modules. To avoid the confusion
we are replacing it with 'private: true' for the private llndk.

'vendor_available: false' will be removed once we update the soong to
read the 'private' property instead of 'vendor_available' for
LLNDK-private libraries.

Bug: 175768895
Test: build
Change-Id: Idf3fe054c59e019f9ed968ca5b6fe49ee6a5e0fc